src/backend/requirements.txt
ffmpeg-python==0.2.0
imageio==2.22.0
imageio-ffmpeg==0.4.7
matplotlib==2.2.2
numpy==1.22.0
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==6.0
scikit-image==0.14.2
scikit-learn==0.19.2
scipy==1.1.0
torch==1.0.0
torchvision==0.2.1
tqdm==4.64.1